backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
databases
httpx
python-dotenv
psycopg2-binary
redis
python-jose
passlib[bcrypt]
python-multipart
boto3
unstructured
langchain
llama-index
elasticsearch
autogen
crewai
